pyproject.toml
[project]
name = "agentic-search"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "langgraph>=0.2",
    "langchain-core>=0.2",
    "anthropic>=0.28",
    "openai>=1.30",
    "elasticsearch[async]>=8.13,<9",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13",
    "tenacity>=8.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "ingest"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires running Elasticsearch",
]
